Fixed lint issues
