+++ src/models/amino/msgs/index.ts
@@ -1,0 +1,9 @@
+// created from 'create-ts-index'
+
+export * from './msg-send'
+export * from './msg-app-stake'
+export * from './msg-app-unstake'
+export * from './msg-app-unjail'
+export * from './msg-node-stake'
+export * from './msg-node-unstake'
+export * from './msg-node-unjail'
