+++ src/models/amino/msgs/msg-node-unjail.ts
@@ -1,0 +1,33 @@
+import { TxMsg } from ".."
+import { typeGuard, validateAddressHex } from "../../.."
+
+/**
+ * Model representing a MsgNodeUnjail to unjail as an Node in the Pocket Network
+ */
+export class MsgNodeUnjail extends TxMsg {
+    public readonly AMINO_KEY: string = "pos/MsgUnjail"
+    public readonly address: string
+
+    /**
+     * @param address {string}
+     */
+    public constructor(address: string) {
+        super()
+        this.address = address
+
+        const errorOrUndefined = validateAddressHex(this.address)
+        if (typeGuard(errorOrUndefined, Error)) {
+            throw errorOrUndefined as Error
+        }
+    }
+
+    public getMsgTypeKey(): string {
+        return this.AMINO_KEY
+    }
+
+    public getMsgValueObj(): {} {
+        return {
+            address: this.address
+        }
+    }
+}
