+++ src/models/amino/msgs/msg-node-unstake.ts
@@ -1,0 +1,33 @@
+import { TxMsg } from ".."
+import { typeGuard, validateAddressHex } from "../../.."
+
+/**
+ * Model representing a MsgNodeStake to unstake as an Node in the Pocket Network
+ */
+export class MsgNodeUnstake extends TxMsg {
+    public readonly AMINO_KEY: string = "pos/MsgBeginUnstake"
+    public readonly nodeAddress: string
+
+    /**
+     * @param nodeAddress {string}
+     */
+    public constructor(nodeAddress: string) {
+        super()
+        this.nodeAddress = nodeAddress
+
+        const errorOrUndefined = validateAddressHex(this.nodeAddress)
+        if (typeGuard(errorOrUndefined, Error)) {
+            throw errorOrUndefined as Error
+        }
+    }
+
+    public getMsgTypeKey(): string {
+        return this.AMINO_KEY
+    }
+
+    public getMsgValueObj(): {} {
+        return {
+            validator_address: this.nodeAddress
+        }
+    }
+}
